Dirty cast to prevent'no matching function for call' error.

+++ src/cyclops/CcdInterface.cpp
@@ -381,9 +381,9 @@ double CcdInterface::profileModel(CyclicCoordinateDescent *ccd, ModelData *model
                       }
                     );
     } else {
-        auto scheduler = TaskScheduler<decltype(boost::make_counting_iterator(bounds.size()))>(
-            boost::make_counting_iterator(0UL),
-            boost::make_counting_iterator(bounds.size()),
+        auto scheduler = TaskScheduler<decltype(boost::make_counting_iterator((int)bounds.size()))>(
+            boost::make_counting_iterator((int)0),
+            boost::make_counting_iterator((int)bounds.size()),
             nThreads);
 
         auto oneTask = [&getBound, &scheduler, &ccdPool, &bounds](unsigned long task) {
